requirements.txt
torch
torchvision
transformers
datasets
peft
evaluat
absl-py
scipy
nltk
rouge-score
